+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Trunk Recorder ChangeLog
 ### Version 3.1.0
 * Updated to the latest version of OP25.
 * Updated P25 Recorder, P25 Trunking and SmartNet Trunking to use the double decimation technique from OP25. It should handle SDRs with a high sample rate better now.
-* Made the Status Socket code an optional add. The WebSocket Library it uses is not longer being maintained and does not build on some systems.
+* Updated to the latest version of the websocketpp library.
 
 ### Version 3.0.1
 * Updated to the latest version of OP25. Supposed performance improvements.